@@ -3,7 +3,8 @@ package com.jiajunhui.xapp.medialoader.callback;
 import android.database.Cursor;
 import android.net.Uri;
 import android.provider.MediaStore;
-import android.support.v4.content.Loader;
+
+import androidx.loader.content.Loader;
 
 import com.jiajunhui.xapp.medialoader.bean.AudioItem;
 import com.jiajunhui.xapp.medialoader.bean.AudioResult;
