@@ -19,9 +19,10 @@ package com.jiajunhui.xapp.medialoader.loader;
 import android.content.Context;
 import android.database.Cursor;
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.LoaderManager;
-import android.support.v4.content.Loader;
+
+import androidx.fragment.app.FragmentActivity;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.Loader;
 
 import com.jiajunhui.xapp.medialoader.callback.OnLoaderCallBack;
 
@@ -58,7 +59,7 @@ public abstract class AbsLoaderCallBack implements LoaderManager.LoaderCallbacks
             if(context!=null){
                 Context ctx = this.context.get();
                 if(ctx!=null){
-                    ((FragmentActivity)ctx).getSupportLoaderManager().destroyLoader(mLoaderId);
+                    LoaderManager.getInstance((FragmentActivity) ctx).destroyLoader(mLoaderId);
                 }
             }
         }catch (Exception e){
